Resolve a weird encoding edge case
So the Mail object knows that its charset is iso-8859-1. But somehow the `body` object thinks it is already encoded in UTF-8, when it is not. Because of this, calling `decoded` will actually convert the mail body to UTF-8. Due to this weird situation, calling `body.decoded` will actually return a string that _says_ it is encoded UTF-8, but it actually is not, and that will cause errors later.

HOW

WHY

[101] pry(#<Mail::Message>):1> charset
=> "iso-8859-1"
[102] pry(#<Mail::Message>):1> body.charset
=> "UTF-8"