create warning post if time after post creation
this is especially cursed. if you create a post inside a `before_action` or `after_action`, you will always cancel the request, because Rails uses "was render called" to know to cancel the request. Rails doesn't care that we used a model to render a template to insert something into the database, we rendered something, and that means the request is going to be cancelled now, no matter what. sigh.